trayendo los cambios de Elifio

+++ src/app/curso/tabla-curso/tabla-curso.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { CursoService } from '../curso.service';
 import { Curso } from '../interface/Curso';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-tabla-curso',
@@ -23,9 +24,27 @@ export class TablaCursoComponent {
   }
 
   eliminar(idcurso:number){
-    this.cursoservice.eliminarCurso(idcurso).subscribe(res=>{
-      this.cursoservice.listarCursos()
-    })
+
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'Esta seguro de eliminar el registro?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si',
+      cancelButtonText:"No"
+    }).then((result) => {
+      if (result.isConfirmed) {
+          this.cursoservice.eliminarCurso(idcurso).subscribe(res=>{
+            Swal.fire('Eliminado!', 'El elemento ha sido eliminado.', 'success');
+            this.cursoservice.listarCursos()
+          }) 
+      }
+    });
+
+
+  
   }
 
 }
