+++ src/app/persona/tabla-persona/tabla-persona.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { PersonaService } from '../persona.service';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-tabla-persona',
@@ -19,9 +20,28 @@ export class TablaPersonaComponent {
   }
 
   eliminar(idpersona:number){
-    this.personaService.eliminarPersona(idpersona).subscribe(res=>{
-      this.personaService.listarPersonas()
-    })
+
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'Esta seguro de eliminar el registro?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si',
+      cancelButtonText:"No"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        // Aquí puedes ejecutar la lógica para eliminar el elemento
+        this.personaService.eliminarPersona(idpersona).subscribe(res=>{
+          Swal.fire('Eliminado!', 'El elemento ha sido eliminado.', 'success');
+          this.personaService.listarPersonas()
+        })
+      }
+    });
+
+
+   
   }
 
   get personas(){
