+++ src/app/seccion/tabla-seccion/tabla-seccion.component.ts
@@ -3,6 +3,7 @@ import { SeccionModule } from '../seccion.module';
 import { Seccion } from '../interface/Seccion';
 import { SeccionService } from '../seccion.service';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-tabla-seccion',
@@ -24,9 +25,27 @@ export class TablaSeccionComponent {
   }
 
   eliminar(idseccion:number){
-    this.seccionService.eliminarSeccion(idseccion).subscribe(res=>{
-      this.seccionService.listarSeccion();
-    })
+
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'Esta seguro de eliminar el registro?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si',
+      cancelButtonText:"No"
+    }).then((result) => {
+      if (result.isConfirmed) {
+          this.seccionService.eliminarSeccion(idseccion).subscribe(res=>{
+            Swal.fire('Eliminado!', 'El elemento ha sido eliminado.', 'success');
+            this.seccionService.listarSeccion();
+          })
+      }
+    });
+
+
+   
   }
