+++ src/app/usuario/tabla-usuario/tabla-usuario.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { UsuarioService } from '../usuario.service';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-tabla-usuario',
@@ -22,9 +23,28 @@ export class TablaUsuarioComponent {
   }
 
   eliminar(idusuario:number){
-    this.usuarioService.eliminarUsuario(idusuario).subscribe(res=>{
-      this.usuarioService.listarUsuarios();
-    })
+
+
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'Esta seguro de eliminar el registro?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si',
+      cancelButtonText:"No"
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+        this.usuarioService.eliminarUsuario(idusuario).subscribe(res=>{
+          Swal.fire('Eliminado!', 'El elemento ha sido eliminado.', 'success');
+          this.usuarioService.listarUsuarios();
+        })
+      }
+    });
+
+   
   }
